Print bibliography section

Given a bibliography file, this function will generate bibliographic entries for one or more types of bib entry.

bibliography_entries(file, startlabel = NULL, endlabel = NULL)

Arguments

file
A path to a bibliography file understood by rmarkdown::pandoc_citeproc_convert().

startlabel
Defunct.

endlabel
Defunct.

Value

A dataset representing the bibliographic entries, suitable for generating a reference section in a document.

Author

Mitchell O'Hara-Wild & Rob J Hyndman

Examples

# Create a bibliography from a set of packages
bib <- tempfile(fileext = ".bib")
knitr::write_bib(c("vitae", "tibble"), bib)

#> Warning: no date field in DESCRIPTION file of package ‘vitae’

# Import the bibliography entries into a CV
bibliography_entries(bib)

#> Warning: running command '/usr/bin/pandoc-citeproc --bib2json /tmp/RtmpPKiJLE/file2a555edc7fbf.bib' had status 1

#> 

#> Error in rmarkdown::pandoc_citeproc_convert(file, type = "json"): Error 1 occurred building shared library.

# The order of these entries can be customised using `dplyr::arrange()`
bibliography_entries(bib) %>%
  arrange(desc(title))

#> Warning: running command '/usr/bin/pandoc-citeproc --bib2json /tmp/RtmpPKiJLE/file2a555edc7fbf.bib' had status 1

#> 

#> Error in rmarkdown::pandoc_citeproc_convert(file, type = "json"): Error 1 occurred building shared library.

# For more complex fields like author, you can also sort by component fields.
# For example, use `author$family` to sort by family names.
bibliography_entries(bib) %>%
  arrange(desc(author$family))

#> Warning: running command '/usr/bin/pandoc-citeproc --bib2json /tmp/RtmpPKiJLE/file2a555edc7fbf.bib' had status 1

#> 

#> Error in rmarkdown::pandoc_citeproc_convert(file, type = "json"): Error 1 occurred building shared library.