Output format for vitae

This output format provides support for including LaTeX dependencies and bibliography entries in extension of the rmarkdown::pdf_document() format.

cv_document(
  ...,
  pandoc_args = NULL,
  pandoc_vars = NULL,
  base_format = rmarkdown::pdf_document
)

Arguments

...

Arguments passed to rmarkdown::pdf_document().

pandoc_args

Additional command line options to pass to pandoc

pandoc_vars

Pandoc variables to be passed to the template.

base_format

An output format function to be used as the base format.