Print bibliography section

Given a bibliography file, this function will generate bibliographic entries for one or more types of bib entry.

bibliography_entries(file, startlabel = NULL, endlabel = NULL)

Arguments

file

A path to a bibliography file understood by rmarkdown::pandoc_citeproc_convert().

startlabel

Defunct.

endlabel

Defunct.

Value

A dataset representing the bibliographic entries, suitable for generating a reference section in a document.

Author

Mitchell O'Hara-Wild & Rob J Hyndman

Examples

# Create a bibliography from a set of packages
bib <- tempfile(fileext = ".bib")
knitr::write_bib(c("vitae", "tibble"), bib)
#> Warning: no date field in DESCRIPTION file of package ‘vitae’

# Import the bibliography entries into a CV
bibliography_entries(bib)
#> # A tibble:   2 × 7
#> # vitae type: bibliography entries
#>   author                            id          issued note  title title…¹ type 
#>   <lst_f_c_>                        <chr>    <csl_dat> <chr> <chr> <chr>   <chr>
#> 1 Kirill Müller, Hadley Wickham     R-tibble      2022 http… Tibb… Tibble  book 
#> 2 Mitchell O’Hara-Wild, Rob Hyndman R-vitae       2023 http… Vita… Vitae   book 
#> # … with abbreviated variable name ¹​`title-short`

# The order of these entries can be customised using `dplyr::arrange()`
bibliography_entries(bib) %>%
  arrange(desc(title))
#> # A tibble:   2 × 7
#> # vitae type: bibliography entries
#>   author                            id          issued note  title title…¹ type 
#> * <lst_f_c_>                        <chr>    <csl_dat> <chr> <chr> <chr>   <chr>
#> 1 Mitchell O’Hara-Wild, Rob Hyndman R-vitae       2023 http… Vita… Vitae   book 
#> 2 Kirill Müller, Hadley Wickham     R-tibble      2022 http… Tibb… Tibble  book 
#> # … with abbreviated variable name ¹​`title-short`

# For more complex fields like author, you can also sort by component fields.
# For example, use `author$family` to sort by family names.
bibliography_entries(bib) %>%
  arrange(desc(author$family))
#> # A tibble:   2 × 7
#> # vitae type: bibliography entries
#>   author                            id          issued note  title title…¹ type 
#> * <lst_f_c_>                        <chr>    <csl_dat> <chr> <chr> <chr>   <chr>
#> 1 Mitchell O’Hara-Wild, Rob Hyndman R-vitae       2023 http… Vita… Vitae   book 
#> 2 Kirill Müller, Hadley Wickham     R-tibble      2022 http… Tibb… Tibble  book 
#> # … with abbreviated variable name ¹​`title-short`