A name variable conforming to the CSL schema

This class provides helper utilities to display, sort, and select attributes from a name in the CSL format.

csl_name(
  family = NULL,
  given = NULL,
  dropping_particle = NULL,
  non_dropping_particle = NULL,
  suffix = NULL,
  comma_suffix = NULL,
  static_ordering = NULL,
  literal = NULL,
  parse_names = NULL
)

csl_names(x = list())

list_of_csl_names(x = list())

Arguments

family

The family name

given

The given name

dropping_particle, non_dropping_particle, suffix, comma_suffix, static_ordering, literal, parse_names

Additional name variable properties as described in the schema.

x

For csl_name(), x should be a list of csl_name(). For list_of_csl_names(), x should be a list of csl_names().

See also